Check if address is IPv4 and only then add it
This fixes problem when ipset restore failed because IPv6 address was
attempted to be added to IPv4 ipset. This commonly resulted in to empty
set on initialization (initial synchronizatio).

This is just workaround. Some addresses can be converted from IPv6 to
IPv4 and IPv6 ipsets support should be added as well.