Add possibility to renew server certificate on client start up
This was originally done in init script but it is much easier to do this
as part of client startup instead. We need this certificate to correctly
authenticate to Sentinel dynamic server. HTTPS is here used to verify
that certificate is really provided by us.

Renew is attempted till it succeeds. The reason is that router might
start without Internet connection but we should be able to on our own
bootstrap connection once it is back online. On the other hand to not do
too many tests when it is clear that router won't be connected to
network any time soon there is also increasing delay between attempts.
This delay is capped at two minutes.